pyproject.toml
[project]
name = "ariasan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "openai>=1.59.2",
    "python-dotenv>=1.0.1",
    "supabase>=2.11.0",
]
